requirements.txt
flask==2.3.3
werkzeug==2.3.7
line-bot-sdk==2.1.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
apscheduler==3.10.1
gunicorn==20.1.0
yfinance==0.2.55
